818903 - Name of the pdf generated for sam system report command should be modified
will now generate PDFs with the name 'katello' or 'headpin'
