+++ cli/src/katello/client/core/system.py
@@ -27,7 +27,7 @@ from katello.client.core.base import BaseAction, Command
 from katello.client.core.utils import test_record, convert_to_mime_type, attachment_file_name, save_report, \
     update_dict_unless_none, stringify_custom_info
 from katello.client.utils.printer import VerboseStrategy
-from katello.client.core.utils import run_spinner_in_bg, wait_for_async_task, SystemAsyncTask, format_date
+from katello.client.core.utils import run_spinner_in_bg, wait_for_async_task, SystemAsyncTask, format_date, get_katello_mode
 from katello.client.utils.encoding import u_str
 from katello.client.utils import printer
 from katello.client.server import ServerRequestError
@@ -759,7 +759,7 @@ class Report(SystemAction):
             report = self.api.report_by_env(environment['id'], convert_to_mime_type(format_in, 'text'))
 
         if format_in == 'pdf':
-            save_report(report[0], attachment_file_name(report[1], 'katello_systems_report.pdf'))
+            save_report(report[0], attachment_file_name(report[1], "%s_systems_report.pdf" % get_katello_mode()))
         else:
             print report[0]
