+++ cli/src/katello/client/main.py
@@ -15,7 +15,7 @@
 # in this software or its documentation.
 #
 
-from katello.client.config import Config
+from katello.client.core.utils import get_katello_mode
 
 from katello.client.core import (
   activation_key,
@@ -329,11 +329,3 @@ def setup_admin(katello_cmd):
     domain_cmd.add_command('update', domain.Update())
     domain_cmd.add_command('delete', domain.Delete())
     katello_cmd.add_command('domain', domain_cmd)
-
-def get_katello_mode():
-    Config()
-    mode = "katello"
-    path = Config.parser.get('server', 'path') or '/katello/api'
-    if "headpin" in path or "sam" in path:
-        mode = "headpin"
-    return mode
